Reformated files, still styling
