+++ client/src/components/Questions/EachAnswer.jsx
@@ -1,24 +1,27 @@
-import React, { useState, useEffect } from 'react';
-import axios from 'axios';
+import React, { useState } from 'react';
 
-export default function IndividualQuestion() {
+export default function EachAnswer() {
   let [helpfulAnswerYes, setHelpfulAnswerYes] = useState(0);
-  const [report, setReport] = useState(0);
   const [answers, setAnswers] = useState([]);
+  const [report, setReport] = useState(0);
 
   function handleAnswerOnClick() {
     setHelpfulAnswerYes(helpfulAnswerYes += 1);
   }
 
   return (
-    <div>
-      <div className="askedquestion">Q:Who what which where why whether how?</div>
-      <div className="answeredquestion">A:One two threee four five six seven</div>
-      <div>
+    <div className="indivanswer">
+      <div className="answerline">
+        <div>A:</div>
+        <div>
+          {answers}
+          {' '}
+          placeholder
+        </div>
+      </div>
+      <div className="helpfula">
         <span>by User1234, Octotober 30th, 2021 |</span>
         <span> | </span>
-      </div>
-      <div>
         <span> Helpful? </span>
         <span onClick={handleAnswerOnClick}>
           Yes
@@ -27,7 +30,7 @@ export default function IndividualQuestion() {
           )
         </span>
         <span> | </span>
-        <span>Add Answer</span>
+        <span>Report</span>
       </div>
     </div>
   );
